NAS-135168: Adapt to changes in crypto plugin
