+++ src/app/pages/credentials/certificates-dash/csr-add/steps/csr-constraints/csr-constraints.component.spec.ts
@@ -4,7 +4,7 @@ import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { ReactiveFormsModule } from '@angular/forms';
 import { createComponentFactory, mockProvider, Spectator } from '@ngneat/spectator/jest';
 import { mockCall, mockApi } from 'app/core/testing/utils/mock-api.utils';
-import { CertificateExtensions } from 'app/interfaces/certificate-authority.interface';
+import { CertificateExtensions } from 'app/interfaces/certificate.interface';
 import { IxFormHarness } from 'app/modules/forms/ix-forms/testing/ix-form.harness';
 import {
   CsrConstraintsComponent,
